refactoring of diff and requirements calculations
